decoupled normalization from qwen specific modules
